requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
pydantic>=2.0.0